Replace teardown wiht afterAll()
