+++ frontend/testing/playwright/playwright.config.ts
@@ -25,22 +25,12 @@ export default defineConfig<ExtendedTestOptions>({
       dependencies: [TestNames.SETUP],
       testDir: './tests/create-app-only/',
       testMatch: '*.spec.ts',
-      teardown: 'teardown-create-app-only',
       use: {
         ...devices['Desktop Chrome'],
         baseURL: process.env.PLAYWRIGHT_TEST_BASE_URL,
         storageState: '.playwright/auth/user.json',
         testAppName: AppNames.CREATE_APP_ONLY,
         headless: true,
-      },
-    },
-    {
-      name: 'teardown-create-app-only',
-      testDir: './tests/create-app-only/',
-      testMatch: '*create-app-only.teardown.ts',
-      use: {
-        baseURL: process.env.PLAYWRIGHT_TEST_BASE_URL,
-        testAppName: AppNames.CREATE_APP_ONLY,
       },
     },
     {
@@ -48,22 +38,12 @@ export default defineConfig<ExtendedTestOptions>({
       dependencies: [TestNames.SETUP],
       testDir: './tests/data-model/',
       testMatch: '*.spec.ts',
-      teardown: 'teardown-data-model',
       use: {
         ...devices['Desktop Chrome'],
         baseURL: process.env.PLAYWRIGHT_TEST_BASE_URL,
         storageState: '.playwright/auth/user.json',
         testAppName: AppNames.DATA_MODEL_APP,
         headless: true,
-      },
-    },
-    {
-      name: 'teardown-data-model',
-      testDir: './tests/data-model/',
-      testMatch: '*data-model.teardown.ts',
-      use: {
-        baseURL: process.env.PLAYWRIGHT_TEST_BASE_URL,
-        testAppName: AppNames.DATA_MODEL_APP,
       },
     },
     {
@@ -71,22 +51,12 @@ export default defineConfig<ExtendedTestOptions>({
       dependencies: [TestNames.SETUP],
       testDir: './tests/dashboard/',
       testMatch: '*.spec.ts',
-      teardown: 'teardown-dashboard',
       use: {
         ...devices['Desktop Chrome'],
         baseURL: process.env.PLAYWRIGHT_TEST_BASE_URL,
         storageState: '.playwright/auth/user.json',
         testAppName: AppNames.DASHBOARD_APP,
         headless: true,
-      },
-    },
-    {
-      name: 'teardown-dashboard',
-      testDir: './tests/dashboard/',
-      testMatch: '*dashboard.teardown.ts',
-      use: {
-        baseURL: process.env.PLAYWRIGHT_TEST_BASE_URL,
-        testAppName: AppNames.DASHBOARD_APP,
       },
     },
     {
@@ -94,22 +64,12 @@ export default defineConfig<ExtendedTestOptions>({
       dependencies: ['setup'],
       testDir: './tests/main-navigation-between-sub-apps/',
       testMatch: '*.spec.ts',
-      teardown: 'teardown-main-navigation-between-sub-apps',
       use: {
         ...devices['Desktop Chrome'],
         baseURL: process.env.PLAYWRIGHT_TEST_BASE_URL,
         storageState: '.playwright/auth/user.json',
         testAppName: AppNames.MAIN_NAVIGATION_APP,
         headless: true,
-      },
-    },
-    {
-      name: 'teardown-main-navigation-between-sub-apps',
-      testDir: './tests/main-navigation-between-sub-apps/',
-      testMatch: '*main-navigation-between-sub-apps.teardown.ts',
-      use: {
-        baseURL: process.env.PLAYWRIGHT_TEST_BASE_URL,
-        testAppName: AppNames.MAIN_NAVIGATION_APP,
       },
     },
     {
