+++ frontend/testing/playwright/tests/create-app-only/create-app-only.spec.ts
@@ -1,6 +1,16 @@
-﻿import { test } from '../../extenders/testExtend';
+﻿import { expect } from '@playwright/test';
+import { Gitea } from '../../helpers/Gitea';
+import { test } from '../../extenders/testExtend';
 import { CreateServicePage } from '../../pages/CreateServicePage';
 import { DashboardPage } from '../../pages/DashboardPage';
+
+test.describe.configure({ mode: 'serial' });
+
+test.afterAll(async ({ request, testAppName }) => {
+  const gitea = new Gitea();
+  const response = await request.delete(gitea.getDeleteAppEndpoint({ app: testAppName }));
+  expect(response.ok()).toBeTruthy();
+});
 
 test('should load dashboard and the user should navigate to create app form', async ({
   page,
