--- frontend/testing/playwright/tests/create-app-only/create-app-only.teardown.ts
@@ -1,9 +1,0 @@
-﻿import { expect } from '@playwright/test';
-import { test } from '../../extenders/testExtend';
-import { Gitea } from '../../helpers/Gitea';
-
-test('should teardown simple-schema-app test', async ({ request, testAppName }) => {
-  const gitea = new Gitea();
-  const response = await request.delete(gitea.getDeleteAppEndpoint({ app: testAppName }));
-  expect(response.ok()).toBeTruthy();
-});
