+++ frontend/testing/playwright/tests/data-model/data-model.spec.ts
@@ -3,6 +3,7 @@ import { test } from '../../extenders/testExtend';
 import { DataModelPage } from '../../pages/DataModelPage';
 import { DesignerApi } from '../../helpers/DesignerApi';
 import type { StorageState } from '../../types/StorageState';
+import { Gitea } from '../../helpers/Gitea';
 
 // This line must be there to ensure that the tests do not run in parallell, and
 // that the before all call is being executed before we start the tests
@@ -13,6 +14,12 @@ test.beforeAll(async ({ testAppName, request, storageState }) => {
   // Create a new app
   const designerApi = new DesignerApi({ app: testAppName });
   const response = await designerApi.createApp(request, storageState as StorageState);
+  expect(response.ok()).toBeTruthy();
+});
+
+test.afterAll(async ({ request, testAppName }) => {
+  const gitea = new Gitea();
+  const response = await request.delete(gitea.getDeleteAppEndpoint({ app: testAppName }));
   expect(response.ok()).toBeTruthy();
 });
